fix: fixed smoke/acceptance test. chore: rename e2e_test to acceptance test.
