+++ .github/workflows/test.yaml
@@ -27,7 +27,7 @@ jobs:
       - name: Test
         run: go test -race -coverprofile=coverage.txt -covermode=atomic
       - name: e2e test
-        run: go test -tags=e2etest
+        run: go test -tags=acctest
         env:
           NEON_API_KEY: ${{ secrets.NEON_API_KEY }}
       - name: Upload coverage to Codecov
